Updates set section card to use card component
